pyproject.toml
[tool.poetry]
name = "Dramatis Personae"
version = "0.9.0"
description = "The DP package"
authors = ["Zaffarelli <zaffarelli@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10.0"
chardet = "==3.0.4"
decorator = "==4.2.1"
django-appconf = "==1.0.3"
django-compressor = "==2.3"
django-extra-views = "==0.12.0"
django-nose = "==1.4.5"
django-sass = "1.1.0"
django-sass-processor = "1.2.2"
django-split-settings = "==0.3.0"
django-test = "==0.4030"
factory-boy = "==2.12.0"
gspread = "==3.6.0"
html5lib = "==1.0.1"
httplib2 = "==0.14.0"
idna = "==2.8"
jdcal = "==1.4.1"
libsass = "==0.19.3"

nose = "==1.3.7"
oauth2client = "==4.1.3"
openpyxl = "==2.5.5"
pyasn1 = "==0.4.7"
pyasn1-modules = "==0.2.7"
python-dateutil = "==2.8.0"
pytz = "==2019.3"
rcssmin = "==1.0.6"
reportlab = "==3.6"
requests = "==2.22.0"
rjsmin = "==1.1.0"
rsa = "==4.0"
split-settings = "==1.0.0"
sqlparse = "==0.3.0"
text-unidecode = "==1.3"
urllib3 = "==1.25.6"
webencodings = "==0.5.1"
xhtml2pdf = "==0.2.4"
Django = "4.0"
et_xmlfile = "==1.0.1"
Faker = "==2.0.3"
Pillow = "7.2.0"
PyYAML = "==5.1"

PyPDF2 = "==1.26.0"
cryptography = "*"
psycopg2-binary = "==2.8.6"
coverage = "*"
pytest = "*"
celery = "==4.4.7"
django-colorfield = "==0.4.1"
certifi = "^2021.10.8"
cffi = "1.14.6"
psycopg2 = "2.8.6"
colour = "^0.1.5"
gunicorn = "^20.1.0"
svglib = "^1.2.1"
CairoSVG = "^2.5.2"
uWSGI = "^2.0.20"
channels = "^4.0.0"
channels-redis = "^4.0.0"
daphne = "^4.0.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
